+++ package.json
@@ -35,6 +35,7 @@
     "file-loader": "^6.0.0",
     "html-webpack-plugin": "^4.3.0",
     "mini-css-extract-plugin": "^0.9.0",
+    "prop-types": "^15.7.2",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
     "sass": "^1.26.10",
     "sass-loader": "^9.0.2",
