+++ src/mock_movies.js
@@ -1,0 +1,46 @@
+const movies_array = [
+    {
+        id: 1, 
+        title: "Avengers", 
+        year: 2015, 
+        genres: ['comedy', 'sci&fi'], 
+        poster: 'https://images-na.ssl-images-amazon.com/images/I/71niXI3lxlL._AC_SY679_.jpg'
+    },
+    {
+        id: 2, 
+        title: "Kill Bill", 
+        year: 2015, 
+        genres: ['drama', 'action'], 
+        poster: 'https://images-na.ssl-images-amazon.com/images/I/61mUJ4Zxc8L._AC_SL1500_.jpg'
+    },
+    {
+        id: 3, 
+        title: "Shawshank's Redemption", 
+        year: 2015, 
+        genres: ['documentary', 'sport', 'triller'], 
+        poster: 'https://www.arthipo.com/image/cache/catalog/genel-tasarim/all-posters/sinema-cinema-film-postersleri/yabanci-filmler/pfilm241-esaretin-bedeli-film-posteri-movie-poster-the-shawshank-redemption_2f7ae7f4-1000x1000.jpg'
+    },
+    {
+        id: 4, 
+        title: "Hulk", 
+        year: 2003, 
+        genres: ['sci&fi','action'], 
+        poster: 'https://images-na.ssl-images-amazon.com/images/I/51heDzhDf8L._AC_.jpg'
+    },
+    {
+        id: 5, 
+        title: "LaLa Land", 
+        year: 2014, 
+        genres: ['musical','romance'], 
+        poster: 'https://m.media-amazon.com/images/I/51CLI6S+Q5L.jpg'
+    },
+    {
+        id: 6, 
+        title: "Inception", 
+        year: 2013, 
+        genres: ['action','sci&fi'], 
+        poster: 'https://images-na.ssl-images-amazon.com/images/I/71wlPWVXb7L._AC_SY500_.jpg'
+    },
+]
+
+export default movies_array;
