+++ src/style.scss
@@ -135,7 +135,7 @@ button {
     &__movie {
         width: 30%;
         min-width: 220px;
-        flex-grow: 1;
+        //flex-grow: 1;
         color: #FFF;
         margin-bottom: 20px;
     }
